requirements.txt
altgraph==0.17.5
bottle==0.13.4
packaging==25.0
proxy_tools==0.1.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pywebview==6.1
setuptools==80.9.0
typing_extensions==4.15.0
